% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{print.scip_control}
\alias{print.scip_control}
\title{Print method for scip_control objects}
\usage{
\method{print}{scip_control}(x, ...)
}
\arguments{
\item{x}{A \code{scip_control} object.}

\item{...}{Ignored.}
}
\value{
Invisible \code{x}.
}
\description{
Print method for scip_control objects
}
