% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{scip_add_sos1_cons}
\alias{scip_add_sos1_cons}
\title{Add a SOS1 constraint to a SCIP model}
\usage{
scip_add_sos1_cons(model, vars, weights = NULL, name = NULL)
}
\arguments{
\item{model}{A SCIP model.}

\item{vars}{Integer vector; 1-based variable indices.}

\item{weights}{Numeric vector; weights determining branching order.}

\item{name}{Character; constraint name.}
}
\value{
Integer; 1-based constraint index.
}
\description{
At most one variable in the set can be nonzero.
}
