% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{scip_add_vars}
\alias{scip_add_vars}
\title{Add multiple variables to a SCIP model}
\usage{
scip_add_vars(model, obj, lb = 0, ub = Inf, vtype = "C", names = NULL)
}
\arguments{
\item{model}{A SCIP model.}

\item{obj}{Numeric vector; objective coefficients.}

\item{lb}{Numeric; lower bounds (scalar or vector). Default \code{0}.}

\item{ub}{Numeric; upper bounds (scalar or vector). Default \code{Inf}.}

\item{vtype}{Character; variable types (scalar or vector). Default \code{"C"}.}

\item{names}{Character vector; variable names. Default auto-generated.}
}
\value{
Integer; 1-based index of first variable added.
}
\description{
Add multiple variables to a SCIP model
}
