% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{scip_get_info}
\alias{scip_get_info}
\title{Get solver information}
\usage{
scip_get_info(model)
}
\arguments{
\item{model}{A SCIP model (after \code{\link{scip_optimize}}).}
}
\value{
A list with \code{solve_time}, \code{nodes}, \code{iterations},
\code{gap}, \code{sol_count}.
}
\description{
Get solver information
}
