% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{scip_get_nsols}
\alias{scip_get_nsols}
\title{Get number of solutions found}
\usage{
scip_get_nsols(model)
}
\arguments{
\item{model}{A SCIP model (after \code{\link{scip_optimize}}).}
}
\value{
Integer.
}
\description{
Get number of solutions found
}
