% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{scip_get_objval}
\alias{scip_get_objval}
\title{Get objective value of best solution}
\usage{
scip_get_objval(model)
}
\arguments{
\item{model}{A SCIP model (after \code{\link{scip_optimize}}).}
}
\value{
Numeric; objective value, or \code{NA} if no solution.
}
\description{
Get objective value of best solution
}
