% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{scip_get_sol}
\alias{scip_get_sol}
\title{Get the k-th solution from the solution pool}
\usage{
scip_get_sol(model, k)
}
\arguments{
\item{model}{A SCIP model.}

\item{k}{Integer; 1-based solution index (1 = best).}
}
\value{
A list with \code{objval} and \code{x}.
}
\description{
Get the k-th solution from the solution pool
}
