% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{scip_get_solution}
\alias{scip_get_solution}
\title{Get the best solution}
\usage{
scip_get_solution(model)
}
\arguments{
\item{model}{A SCIP model (after \code{\link{scip_optimize}}).}
}
\value{
A list with \code{objval} and \code{x}.
}
\description{
Get the best solution
}
