% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{scip_model_free}
\alias{scip_model_free}
\title{Free a SCIP model}
\usage{
scip_model_free(model)
}
\arguments{
\item{model}{A SCIP model.}
}
\value{
Invisible \code{NULL}.
}
\description{
Explicitly frees the SCIP model and all associated memory.
The model is also freed automatically when garbage collected.
}
