% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{scip_optimize}
\alias{scip_optimize}
\title{Solve a SCIP model}
\usage{
scip_optimize(model)
}
\arguments{
\item{model}{A SCIP model.}
}
\value{
Invisible \code{NULL}. Use \code{\link{scip_get_status}} and
\code{\link{scip_get_solution}} to retrieve results.
}
\description{
Solve a SCIP model
}
