% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{scip_set_objective_sense}
\alias{scip_set_objective_sense}
\title{Set objective sense}
\usage{
scip_set_objective_sense(model, sense = "minimize")
}
\arguments{
\item{model}{A SCIP model.}

\item{sense}{Character; \code{"minimize"} (default) or \code{"maximize"}.}
}
\value{
Invisible \code{NULL}.
}
\description{
Set objective sense
}
